Move around filesystem references and files.
