+++ lib/celluloid/groups/spawner.rb
