+++ spec/celluloid/groups/spawner_spec.rb
