+++ spec/celluloid/tasks/fibered_spec.rb
