+++ spec/celluloid/tasks/threaded_spec.rb
